fix(plugins): refuse staged publication after startup cancellation
